samples/cloudrun/cloudrun/sqlserver/requirements.txt
cloud-sql-python-connector[pytds]
sqlalchemy-pytds
google-cloud-secret-manager
Flask
gunicorn
